Adjust placement of 'X_normalization' check
Moved placement of 'X_normalization' check to where cxg_adata.raw is set. Flattener should now check for mfinal_obj['X_normalization'] and if False will skip setting cxg_adata.raw and instead set raw to be in place of 'normalized'. Then will also set 'feature_is_filtered' to False to coincide with this change.
If mfinal_obj['X_normalization'] is True, the cxg_adata.raw will be set to cxg_adata_raw as normal.

Had to include a check for 'X_normalization' in the quality_check function, as before it was checking if var index was longer than raw var index, which would error because if 'X_normalization' is False then there is no raw var index.