Fix for new error found from previous ticket
This commit falls outside the scope of TOOLS-144, but is instead fixing a newly found error on a commit from TOOLS-138_mapped_reference_annotation_should_be_from_proc_matrix_file.

Replacing if statement at ln 1275 (which is for checking the presence of 'genome_annotations' in the processedmatrixfile) with the more concise and robust .get and return None if not present method. This is to solve an error that occurs when 'genome_annotations' is not present.